+++ packages/suspense-website/src/examples/createIntervalCache/hook.ts
@@ -1,6 +1,7 @@
 import { sourceCodeCache } from "./cache";
 
 // REMOVE_BEFORE
+
 import { useIntervalCacheStatus } from "suspense";
 
 function StatusBadge({