+++ packages/suspense/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.0.28
+* Add `useImperativeIntervalCacheValue` hook for imperatively loading and subscribing to `createIntervalCache` data.
+* Add `getValue` and `getValueIfCached` methods to `createIntervalCache` type.
+
 ## 0.0.27
 * Removed `useWeakRef` config option for `createCache` and replaced with `getCache` option in order to support LRU type caches in addition to `WeakRef` based caches. Thanks to @cevr for contributing to this release!